% Generated by roxygen2 (4.0.1): do not edit by hand
\name{caDescribeOpts}
\alias{caDescribeOpts}
\title{A function for gathering all the description options}
\usage{
caDescribeOpts(show_tot_perc = FALSE, numb_first = TRUE,
  continuous_fn = describeMean, prop_fn = describeProp,
  factor_fn = describeFactors, show_missing = FALSE, digits = 1,
  colnames = c("Total", "Event"))
}
\arguments{
\item{show_tot_perc}{Show percentages for the total column}

\item{numb_first}{Whether to show the number before the percentages}

\item{continuous_fn}{Stat function used for the descriptive statistics,
defaults to \code{\link{describeMean}}}

\item{prop_fn}{Stat function used for the descriptive statistics,
defaults to \code{\link{describeProp}}}

\item{factor_fn}{Stat function used for the descriptive statistics,
defaults to \code{\link{describeFactors}}}

\item{show_missing}{Show missing variables in the descriptive columns}

\item{digits}{Number of digits to use in the descriptive columns.
Defaults to the general digits if not specified.}

\item{colnames}{The names of the two descriptive columns. By default
Total and Event.}
}
\value{
\code{list} Returns a list with all the options
}
\description{
Since there are so many different description options
for the \code{\link{printCrudeAndAdjusted}} function they
have been gathered into a list. This function is simply a
helper in order to generate a valid list.
}

