% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Greg}
\alias{Greg}
\alias{Greg-package}
\title{A package for regression handling}
\description{
This R-package provides additional functions that help you
 convey regression models. It has also a few functions for
 handling robust confidence intervals for the ols() regression
 in the rms-package. This package is closely related to the Gmisc-package.
}

