% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms_SandwichAddon.R
\name{model.matrix.ols}
\alias{model.matrix.ols}
\title{A fix for the model.matrix}
\usage{
\method{model.matrix}{ols}(object, ...)
}
\arguments{
\item{object}{A Model}

\item{...}{Parameters passed on}
}
\value{
matrix
}
\description{
The \code{\link[stats]{model.matrix.lm}()} that the \code{\link[rms]{ols}} falls back upon
"forgets" the intercept value and behaves unreliable in
the \code{\link[sandwich]{vcovHC}()} funcitons. I've therefore created this subfunction
to generate the actual \code{\link[stats]{model.matrix}()} by just accessing the formula.
}
\keyword{internal}
