% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prCaAddUserReferences}
\alias{prCaAddUserReferences}
\title{Adds references}
\usage{
prCaAddUserReferences(reordered_groups, var_order, add_references,
  add_references_pos, reference_zero_effect)
}
\arguments{
\item{reordered_groups}{The value matrix that needs refrences}

\item{var_order}{The output from the \code{\link{prMapVariable2Name}}}
}
\value{
\code{matrix} The \code{reordered_groups} with references and the
 attribute "var_order" in order to keep track of no. of variables per row.
}
\description{
Adds references
}
\seealso{
Other printCrudeAndAdjusted functions: \code{\link{latex.printCrudeAndAdjusted}},
  \code{\link{print.printCrudeAndAdjusted}},
  \code{\link{printCrudeAndAdjustedModel}};
  \code{\link{prCaAddRefAndStat}};
  \code{\link{prCaAddReference}};
  \code{\link{prCaGetImputationCols}};
  \code{\link{prCaGetOrderVariables}};
  \code{\link{prCaGetRowname}};
  \code{\link{prCaGetVnStats}};
  \code{\link{prCaPrepareCrudeAndAdjusted}};
  \code{\link{prCaReorderReferenceDescribe}};
  \code{\link{prCaReorder}}; \code{\link{prCaSetRownames}}
}
\keyword{internal}

