% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prCaGetVnStats}
\alias{prCaGetVnStats}
\title{Gets the variable stats}
\usage{
prCaGetVnStats(model, vn, outcome, ds, add_references, add_references_pos,
  desc_args)
}
\arguments{
\item{model}{The model}

\item{vn}{The variable name}

\item{outcome}{The outcome vector}

\item{ds}{The dataset}
}
\value{
\code{matrix} A matrix from \code{\link[Gmisc]{getDescriptionStatsBy}} or
 \code{\link{prGetStatistics}}
}
\description{
Gets the variable stats
}
\seealso{
Other printCrudeAndAdjusted functions: \code{\link{latex.printCrudeAndAdjusted}},
  \code{\link{print.printCrudeAndAdjusted}},
  \code{\link{printCrudeAndAdjustedModel}};
  \code{\link{prCaAddRefAndStat}};
  \code{\link{prCaAddReference}};
  \code{\link{prCaAddUserReferences}};
  \code{\link{prCaGetImputationCols}};
  \code{\link{prCaGetRowname}};
  \code{\link{prCaPrepareCrudeAndAdjusted}};
  \code{\link{prCaReorderReferenceDescribe}};
  \code{\link{prCaReorder}};
  \code{\link{prCaSelectAndOrderVars}};
  \code{\link{prCaSetRownames}}
}
\keyword{internal}

