% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private_printCaA.R
\name{prCaPrepareCrudeAndAdjusted}
\alias{prCaPrepareCrudeAndAdjusted}
\title{Prettify the text}
\usage{
prCaPrepareCrudeAndAdjusted(x, ci_lim, digits, sprintf_ci_str)
}
\arguments{
\item{x}{The value matrix from getCrudeAndAdjusted}

\item{ci_lim}{The limits of the confidence interval}

\item{digits}{The number of decimal digits to use}

\item{sprintf_ci_str}{The \code{\link{sprintf}} code for the confidence interval}
}
\value{
\code{matrix} A string matrix with the values formated
}
\description{
Sets the number of digits, formats the confidence interval and
changes the number of cols into 4 where the upper and lower CI
meet in one string column
}
\seealso{
Other printCrudeAndAdjusted functions: \code{\link{latex.printCrudeAndAdjusted}},
  \code{\link{print.printCrudeAndAdjusted}},
  \code{\link{printCrudeAndAdjustedModel}},
  \code{\link{rbind.printCrudeAndAdjusted}};
  \code{\link{prCaAddRefAndStat}};
  \code{\link{prCaAddReference}};
  \code{\link{prCaAddUserReferences}};
  \code{\link{prCaGetImputationCols}};
  \code{\link{prCaGetRowname}};
  \code{\link{prCaGetVnStats}};
  \code{\link{prCaReorderReferenceDescribe}};
  \code{\link{prCaReorder}};
  \code{\link{prCaSelectAndOrderVars}};
  \code{\link{prCaSetRownames}}
}
\keyword{internal}

