% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prCaReorderReferenceDescribe}
\alias{prCaReorderReferenceDescribe}
\title{Adds the ordering, references, and descriptions}
\usage{
prCaReorderReferenceDescribe(x, model, order, var_order, add_references,
  add_references_pos, reference_zero_effect, ds, desc_column, desc_args,
  use_labels)
}
\arguments{
\item{x}{The main value matrix from the \code{\link{prCaPrepareCrudeAndAdjusted}}}

\item{model}{The model}

\item{order}{A vector A vector with regular expressions for each group.}

\item{var_order}{The output from the \code{\link{prMapVariable2Name}}}

\item{ds}{The dataset from the model}
}
\value{
The reordered groups as a matrix
}
\description{
This is a wrapper function around some more basic functions that
\code{\link{printCrudeAndAdjusted uses}}.
}
\seealso{
Other printCrudeAndAdjusted functions: \code{\link{latex.printCrudeAndAdjusted}},
  \code{\link{print.printCrudeAndAdjusted}},
  \code{\link{printCrudeAndAdjustedModel}};
  \code{\link{prCaAddRefAndStat}};
  \code{\link{prCaAddReference}};
  \code{\link{prCaAddUserReferences}};
  \code{\link{prCaGetImputationCols}};
  \code{\link{prCaGetOrderVariables}};
  \code{\link{prCaGetRowname}};
  \code{\link{prCaGetVnStats}};
  \code{\link{prCaPrepareCrudeAndAdjusted}};
  \code{\link{prCaReorder}}; \code{\link{prCaSetRownames}}
}
\keyword{internal}

