% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private_printCaA.R
\name{prCaSelectAndOrderVars}
\alias{prCaSelectAndOrderVars}
\title{Re-order variables}
\usage{
prCaSelectAndOrderVars(names, order, ok2skip = FALSE)
}
\arguments{
\item{names}{The names of the variables}

\item{order}{The order regular expression}

\item{ok2skip}{If you have the intercept then
 it should be ok for the function to skip that
 variable if it isn't found among the variable list}
}
\value{
\code{vector} A vector containing the greps
}
\description{
Re-order variables
}
\seealso{
Other printCrudeAndAdjusted functions: \code{\link{knit_print.printCrudeAndAdjusted}},
  \code{\link{latex.printCrudeAndAdjusted}},
  \code{\link{print.printCrudeAndAdjusted}},
  \code{\link{printCrudeAndAdjustedModel}},
  \code{\link{rbind.printCrudeAndAdjusted}};
  \code{\link{prCaAddRefAndStat}};
  \code{\link{prCaAddReference}};
  \code{\link{prCaAddUserReferences}};
  \code{\link{prCaGetImputationCols}};
  \code{\link{prCaGetRowname}};
  \code{\link{prCaGetVnStats}};
  \code{\link{prCaPrepareCrudeAndAdjusted}};
  \code{\link{prCaReorderReferenceDescribe}};
  \code{\link{prCaReorder}}; \code{\link{prCaSetRownames}}
}
\keyword{internal}
\keyword{intrenal}

