% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCrudeAndAdjustedModel.R
\name{prClearPCAclass}
\alias{prClearPCAclass}
\title{Removes the printCrudeAndAdjusted class from arguments}
\usage{
prClearPCAclass(pca)
}
\arguments{
\item{...}{The parameters to the cbind/rbind functions}
}
\value{
list
}
\description{
Removes the printCrudeAndAdjusted class from arguments
}
\keyword{internal}

