% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prConvertShowMissing}
\alias{prConvertShowMissing}
\title{A functuon for converting a useNA variable}
\usage{
prConvertShowMissing(useNA)
}
\arguments{
\item{useNA}{Boolean or "no", "ifany", "always"}
}
\value{
string
}
\description{
The variable is suppose to be directly compatible with
table(..., useNA=useNA). It throughs an error
if not compatible
}
\keyword{internal}

