% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prExtractOutcomeFromModel}
\alias{prExtractOutcomeFromModel}
\title{Get model outcome}
\usage{
prExtractOutcomeFromModel(model)
}
\arguments{
\item{model}{The fitted model.1}
}
\value{
vector
}
\description{
Uses the model to extract the outcome variable. Throws
error if unable to find the outcome.
}
\keyword{internal}

