% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prExtractPredictorsFromModel}
\alias{prExtractPredictorsFromModel}
\title{Get model predictors}
\usage{
prExtractPredictorsFromModel(model)
}
\arguments{
\item{model}{The fitted model.}
}
\value{
data.frame
}
\description{
Uses the model to extract the dataset. Throws
error if unable to find the data
}
\author{
max
}

