% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prFindRownameMatches}
\alias{prFindRownameMatches}
\title{Looks for unique rowname match without grep}
\usage{
prFindRownameMatches(rnames, vn, vars)
}
\arguments{
\item{rnames}{A vector with the rownames that are looked for}

\item{vn}{The variable name that is of interest}

\item{vars}{A vector with all the names and the potentially competing names}
}
\value{
integer A vector containing the position of the matches

TODO: remove this function in favor of the more powerful prMapVariable2Name
}
\description{
Since a rowname may contain characters reserved by regular
expressions I've found it easier to deal with the rowname
finding by just checking for matching strings at the beginning
of the name while at the same time excluding names that have the
same stem, i.e. DM and DM_COMP will cause an issue since DM will
match both rows.
}
\keyword{internal}

