% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prMapVariable2Name}
\alias{prMapVariable2Name}
\title{A function that tries to resolve what variable corresponds to what row}
\usage{
prMapVariable2Name(var_names, available_names, data)
}
\arguments{
\item{var_names}{The variable names that are saught after}

\item{available_names}{The names that are available to search through}

\item{data}{The data set that is sught after}
}
\value{
\code{list} Returns a list with each element has the corresponding
 variable name and a subsequent list with the parameters \code{no_rows}
 and \code{location} indiciting the number of rows corresponding to that
 element and where those rows are located. For factors the list also contains
 \code{lvls} and \code{no_lvls}.
}
\description{
As both the \code{\link{getCrudeAndAdjustedModelData}} and the
\code{\link{printCrudeAndAdjustedModel}} need to now exactly
what name from the \code{\link[stats]{coef}}/\code{\link[rms]{summary}}
correspond to we for generalizeability this rather elaborate function.
}

