% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_plotHR.R
\name{prPhConfIntPlot}
\alias{prPhConfIntPlot}
\title{Plots the confidence intervals}
\usage{
prPhConfIntPlot(model_data, color, polygon, lwd, lty)
}
\arguments{
\item{model_data}{A data frame with 'xvalues', 'upper', and 'lower'
columns.}

\item{color}{The color of the line/polygon}

\item{polygon}{Boolean indicating polygon or line}

\item{lwd}{Line width - see \code{\link[grid]{gpar}}}

\item{lty}{Line type - see \code{\link[grid]{gpar}}}
}
\value{
\code{void} The function performs the print
}
\description{
Uses \code{\link[graphics]{polygon}} or 
\code{\link[graphics]{lines}} to plot confidence
intervals.
}
\keyword{internal}
