% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_plotHR.R
\name{prPhDensityPlot}
\alias{prPhDensityPlot}
\title{Plot a density on the datapoints}
\usage{
prPhDensityPlot(xvalues, color)
}
\arguments{
\item{xvalues}{The xvalues that are used for the density}

\item{color}{The color of the density polygon}
}
\value{
\code{void}
}
\description{
Plot a density on the datapoints
}
\keyword{internal}

