% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prPhEstimate}
\alias{prPhEstimate}
\title{Gets the non-linear function's estimate}
\usage{
prPhEstimate(model, term.label, ylog, cntrst, xlim, alpha, new_data)
}
\arguments{
\item{model}{The fit of the model to be plotted}

\item{term.label}{The name of the label}

\item{ylog}{If the outcome should be presented in the anti-log form, i.e.
\code{exp()}.}

\item{cntrst}{A boolean that indicates if the \code{\link[rms]{contrast}}
function is to be deployed for \pkg{rms} generated functions. The nice
thing is that you get the median as a reference by default.}

\item{xlim}{The xlim if provided}

\item{new_data}{If not provided the function looks for the most common values
i.e. median for continuous and mode for factors.}
}
\value{
\code{data.frame} with the columns xvalues, fit, ucl, lcl
}
\description{
The function uses predict if not specified contrast
in order to attain the estimate, upper, and lower
confidence interval
}
\keyword{internal}

