% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_plotHR.R
\name{prPhNewData}
\alias{prPhNewData}
\title{A function for retrieving new_data argument for predict}
\usage{
prPhNewData(model, term.label, xlim)
}
\arguments{
\item{model}{The model fit from \code{\link[survival]{coxph}()}
or \code{\link[rms]{cph}()}}

\item{term.label}{The label that is the one that \code{\link{plotHR}()}
intends to plot.}

\item{xlim}{The x-limits for the plot if any}
}
\value{
\code{data.frame}
}
\description{
A function for retrieving new_data argument for predict
}
\keyword{internal}
