% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addNonlinearity.R
\name{addNonlinearity}
\alias{addNonlinearity}
\alias{addNonlinearity.default}
\alias{addNonlinearity.rms}
\title{Add a nonlinear function to the model}
\usage{
addNonlinearity(model, variable, spline_fn, flex_param = 2:7, min_fn = AIC,
  sig_level = 0.05, verbal = FALSE, workers, ...)

\method{addNonlinearity}{default}(model, variable, spline_fn,
  flex_param = 2:7, min_fn = AIC, sig_level = 0.05, verbal = FALSE,
  workers, ...)

\method{addNonlinearity}{rms}(model, variable, spline_fn, flex_param = 2:7,
  min_fn = AIC, sig_level = 0.05, verbal = FALSE, workers, ...)
}
\arguments{
\item{model}{The model that is to be evaluated and adatpted for non-linearity}

\item{spline_fn}{Either a string or a function that is to be used
for testing alternative non-linearity models}

\item{flex_param}{A \code{vector} with values that are to be tested as the
default second parameter for the non-linearity function that you want to
evaluate. This defaults to 2:7, for the \code{\link[splines]{ns}} it tests
the degrees of freedom ranging between 2 and 7.}

\item{min_fn}{This is the function that we want to minmized if the variable supports
the non-linearity assumption. E.g. \code{\link[stats]{BIC}} or
\code{\link[stats]{AIC}}, note that the \code{BIC} will in the majority of cases
support a lower complexity than the \code{AIC}.}

\item{sig_level}{The significance level for which the non-linearity is deemed
as significant, defaults to 0.05.}

\item{verbal}{Set this to \code{TRUE} if you want print statements with the
anova test and the chosen knots.}

\item{workers}{The function tries to run everything in parallel. Under some
 circumstances you may want to restrict the number of parallel threads to less
 than the defaul \code{dectectCores() - 1}, e.g. you may run out of memory
 then you can provide this parameter. If you do not want to use parallel then
 simply set workers to \code{FALSE}.}

\item{param}{The name of the parameter that is to be tested for non-linearity.
\emph{Note} that the variable should be included plain (i.e. as a linear variable)
form in the model.}
}
\description{
This function takes a model and adds a non-linear function if
the likelihood-ratio supports this (via the
\code{\link[stats]{anova}(..., test="chisq")} test for \pkg{stats}
while for \pkg{rms} you need to use the \code{rcs} spline
that is automatically evaluated for non-linearity).
}
\keyword{internal}

