% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplotRegrObj.R
\name{forestplotRegrObj}
\alias{forestplotRegrObj}
\alias{forestplotRegrObj.default}
\alias{forestplotRegrObj.coxph}
\alias{forestplotRegrObj.lrm}
\alias{forestplotRegrObj.lm}
\alias{forestplotRegrObj.glm}
\alias{forestplotRegrObj.list}
\alias{fpBoxSize}
\title{Forest plot for multiple models}
\usage{
forestplotRegrObj(
  regr.obj,
  boxsize,
  postprocess_estimates.fn = function(x) x,
  xlab,
  xlog,
  exp,
  estimate.txt = xlab,
  rowname = "Variable",
  ci.txt = "CI",
  ci.glue = "{lower} to {higher}",
  digits = 1,
  zero,
  get_box_size = fpBoxSize,
  ...
)

\method{forestplotRegrObj}{default}(
  regr.obj,
  skip.variables,
  add.empty_row,
  order.regexps,
  order.addrows,
  boxsize,
  rowname.fn,
  xlab,
  xlog,
  exp,
  estimate.txt = xlab,
  zero,
  get_box_size = fpBoxSize,
  ...
)

\method{forestplotRegrObj}{coxph}(
  regr.obj,
  xlab = "Hazard Ratio",
  estimate.txt = "HR",
  xlog = TRUE,
  zero = 1,
  exp = TRUE,
  ...
)

\method{forestplotRegrObj}{lrm}(
  regr.obj,
  xlab = "Odds ratio",
  estimate.txt = "HR",
  xlog = TRUE,
  zero = 1,
  exp = TRUE,
  ...
)

\method{forestplotRegrObj}{lm}(
  regr.obj,
  xlab = "Effect",
  estimate.txt = "Coef",
  xlog = FALSE,
  zero = 0,
  exp = FALSE,
  ...
)

\method{forestplotRegrObj}{glm}(
  regr.obj,
  xlab = NULL,
  xlog = NULL,
  zero = NULL,
  estimate.txt = NULL,
  exp = NULL,
  ...
)

\method{forestplotRegrObj}{list}(
  regr.obj,
  xlab = NULL,
  xlog = NULL,
  zero = NULL,
  estimate.txt = NULL,
  exp = NULL,
  ...
)

fpBoxSize(p_values, variable_count, boxsize, significant = 0.05)
}
\arguments{
\item{regr.obj}{A regression model object. It should be of coxph, crr or glm class.
Warning: The glm is not fully tested.}

\item{boxsize}{The default box size}

\item{postprocess_estimates.fn}{A function that takes the regression outputs and returns
the same data with modifications. The input columns are:

* `Rowname`
* `Coef`
* `Lower`
* `Upper`
* `Sort`}

\item{xlab}{x-axis label}

\item{xlog}{If TRUE, x-axis tick marks are to follow a logarithmic scale, e.g. for
logistic regression (OR), survival estimates (HR), Poisson regression etc.
\emph{Note:} This is an intentional break with the original \code{forestplot}
function as I've found that exponentiated ticks/clips/zero effect are more
difficult to for non-statisticians and there are sometimes issues with rounding
the tick marks properly.}

\item{exp}{Report in exponential form. Default true since the function was built for
use with survival models.}

\item{estimate.txt}{The text above the estimate, e.g. Est, HR}

\item{rowname}{The name of the variables}

\item{ci.txt}{The text above the confidence interval, defaults to `"CI"`}

\item{ci.glue}{The string used for [glue::glue()] the `lower` and `higher`
confidence intervals together.}

\item{digits}{The number of digits to round presented values to}

\item{zero}{Indicates what is zero effect. For survival/logistic fits the zero is
1 while in most other cases it's 0.}

\item{get_box_size}{A function for extracting the box sizes}

\item{...}{Passed to \code{\link[forestplot]{forestplot}()}}

\item{rowname.fn}{A function that takes a row name and sees if it needs
beautifying. The function has only one parameter the coefficients name and should
return a string or expression.}

\item{p_values}{The p-values that will work as the foundation for the box size}

\item{variable_count}{The number of variables}

\item{significant}{Level of significance .05}
}
\description{
Plot different model fits with similar variables in order to
compare the model's estimates and confidence intervals. Each
model is represented by a separate line on top of eachother
and are therefore ideal for comparing different models. This
extra appealing when you have lots of variables included in
the models.
}
\examples{
org.par <- par("ask" = TRUE)

# simulated data to test
set.seed(10)
ftime <- rexp(200)
fstatus <- sample(0:2, 200, replace = TRUE)
cov <- data.frame(
  x1 = runif(200),
  x2 = runif(200),
  x3 = runif(200)
)

library(rms)
dd <- datadist(cov)
options(datadist = "dd")

fit1 <- cph(Surv(ftime, fstatus == 1) ~ x1 + x2 + x3, data = cov)
fit2 <- cph(Surv(ftime, fstatus == 2) ~ x1 + x2 + x3, data = cov)

forestplotRegrObj(regr.obj = fit1, new_page = TRUE)

library(forestplot)
forestplotRegrObj(
  regr.obj = list(fit1, fit2),
  legend = c("Status = 1", "Status = 2"),
  legend_args = fpLegend(title = "Type of regression"),
  new_page = TRUE
)

modifyNameFunction <- function(x) {
  if (x == "x1") {
    return("Covariate A")
  }

  if (x == "x2") {
    return(expression(paste("My ", beta[2])))
  }

  return(x)
}

forestplotRegrObj(
  regr.obj = list(fit1, fit2),
  col = fpColors(box = c("darkblue", "darkred")),
  variablesOfInterest.regexp = "(x2|x3)",
  legend = c("First model", "Second model"),
  legend_args = fpLegend(title = "Models"),
  rowname.fn = modifyNameFunction, new_page = TRUE
)

par(org.par)
}
\seealso{
Other forestplot wrappers: 
\code{\link{forestplotCombineRegrObj}()}
}
\concept{forestplot wrappers}
