% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prCaAddReference}
\alias{prCaAddReference}
\title{Adds a reference to value matrix}
\usage{
prCaAddReference(vn, var_order, values, add_references_pos,
  reference_zero_effect, ds, use_labels)
}
\arguments{
\item{vn}{Variable name}

\item{var_order}{The output from the \code{\link{prMapVariable2Name}}}

\item{values}{The value matrix}

\item{add_references_pos}{The position within the factors if it exists}

\item{reference_zero_effect}{The reference zero effect}

\item{ds}{The data set}

\item{use_labels}{If labels should be used for row names}
}
\value{
\code{matrix} A matrix with rgroup and n.rgroup attributes
}
\description{
Adds a reference to value matrix
}
\author{
max
}

