% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prCaAddReference}
\alias{prCaAddReference}
\title{Adds a reference to value matrix}
\usage{
prCaAddReference(vn, var_order, values, add_references_pos,
  reference_zero_effect, ds, use_labels)
}
\arguments{
\item{vn}{Variable name}

\item{var_order}{The output from the \code{\link{prMapVariable2Name}}}

\item{values}{The value matrix}

\item{ds}{The data set}
}
\value{
\code{matrix} A matrix with rgroup and n.rgroup attributes
}
\description{
Adds a reference to value matrix
}
\seealso{
Other printCrudeAndAdjusted functions: \code{\link{latex.printCrudeAndAdjusted}},
  \code{\link{print.printCrudeAndAdjusted}},
  \code{\link{printCrudeAndAdjustedModel}};
  \code{\link{prCaAddRefAndStat}};
  \code{\link{prCaAddUserReferences}};
  \code{\link{prCaGetImputationCols}};
  \code{\link{prCaGetRowname}};
  \code{\link{prCaGetVnStats}};
  \code{\link{prCaPrepareCrudeAndAdjusted}};
  \code{\link{prCaReorderReferenceDescribe}};
  \code{\link{prCaReorder}};
  \code{\link{prCaSelectAndOrderVars}};
  \code{\link{prCaSetRownames}}
}
\keyword{internal}

