% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private_printCaA.R
\name{prCaGetImputationCols}
\alias{prCaGetImputationCols}
\title{Function for retrieving the imputation arguments}
\usage{
prCaGetImputationCols(impute_args, output_mtrx, model, data)
}
\arguments{
\item{impute_args}{The imputation arguments from \code{\link{printCrudeAndAdjusted}}
function call.}

\item{output_mtrx}{The reordered groups matrix (a nx4 matrix)
that have been prepared in for the \code{\link{printCrudeAndAdjusted}}
function. It is important that the references
if any have been added.}

\item{model}{The imputation model. Currently only \code{\link[Hmisc]{fit.mult.impute}}
is supported by the function.}

\item{data}{The data that has been used for generating the model.}
}
\value{
\code{matrix} Returns a matrix with the requested columns
}
\description{
Function for retrieving the imputation arguments
}
\seealso{
Other printCrudeAndAdjusted functions: \code{\link{latex.printCrudeAndAdjusted}},
  \code{\link{print.printCrudeAndAdjusted}},
  \code{\link{printCrudeAndAdjustedModel}};
  \code{\link{prCaAddRefAndStat}};
  \code{\link{prCaAddReference}};
  \code{\link{prCaAddUserReferences}};
  \code{\link{prCaGetRowname}};
  \code{\link{prCaGetVnStats}};
  \code{\link{prCaPrepareCrudeAndAdjusted}};
  \code{\link{prCaReorderReferenceDescribe}};
  \code{\link{prCaReorder}};
  \code{\link{prCaSelectAndOrderVars}};
  \code{\link{prCaSetRownames}}
}
\keyword{internal}

