% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private_printCaA.R
\name{prCaGetVnStats}
\alias{prCaGetVnStats}
\title{Gets the variable stats}
\usage{
prCaGetVnStats(model, vn, outcome, ds, add_references, add_references_pos,
  desc_args)
}
\arguments{
\item{model}{The model}

\item{vn}{The variable name}

\item{outcome}{The outcome vector}

\item{ds}{The dataset}

\item{add_references}{True if it should use the data set to look for
references, otherwise supply the function with a vector with names.
Sometimes you want to indicate the reference row for each group.
This needs to be just as many as the  groups as the order identified.
Use NA if you don't want to have a reference for that particular group.}

\item{add_references_pos}{The position where a reference should be added.
Sometimes you don't want the reference to be at the top, for instance
if you have age groups then you may have < 25, 25-39, 40-55, > 55 and
you have the reference to be 25-39 then you should set the reference
list for \code{age_groups} as \code{add_references_pos = list(age_groups = 2)}
so that you have the second group as the position for the reference.}

\item{desc_args}{The description arguments that are to be used for the
the description columns. The options/arguments should be generated by the
\code{\link{caDescribeOpts}} function.}
}
\value{
\code{matrix} A matrix from \code{\link[Gmisc]{getDescriptionStatsBy}} or
 \code{\link{prGetStatistics}}
}
\description{
Gets the variable stats
}
\seealso{
Other printCrudeAndAdjusted functions: \code{\link{knit_print.printCrudeAndAdjusted}},
  \code{\link{latex.printCrudeAndAdjusted}},
  \code{\link{print.printCrudeAndAdjusted}},
  \code{\link{printCrudeAndAdjustedModel}},
  \code{\link{rbind.printCrudeAndAdjusted}};
  \code{\link{prCaAddRefAndStat}};
  \code{\link{prCaAddReference}};
  \code{\link{prCaAddUserReferences}};
  \code{\link{prCaGetImputationCols}};
  \code{\link{prCaGetRowname}};
  \code{\link{prCaPrepareCrudeAndAdjusted}};
  \code{\link{prCaReorderReferenceDescribe}};
  \code{\link{prCaReorder}};
  \code{\link{prCaSelectAndOrderVars}};
  \code{\link{prCaSetRownames}}
}
\keyword{internal}

