% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prEnvModelCall}
\alias{prEnvModelCall}
\title{Runs an \code{do.call} within the environment of the model}
\usage{
prEnvModelCall(model, what, ...)
}
\arguments{
\item{model}{The model used}

\item{what}{The function or non-empty character string used for
\code{\link[base]{do.call}}}

\item{...}{Additional arguments passed to the function}
}
\description{
Sometimes the function can't find some of the variables that
were available when running the original variable. This function
uses the \code{\link[stats]{as.formula}} together with
\code{\link[base]{environment}} in order to get the environment
that the original code used.
}
\keyword{internal}

