% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prExtractOutcomeFromModel}
\alias{prExtractOutcomeFromModel}
\title{Get model outcome}
\usage{
prExtractOutcomeFromModel(model, mf)
}
\arguments{
\item{model}{The fitted model}

\item{mf}{The dataset that the model is fitted to - if missing it
uses the \code{\link[stats]{model.frame}} dataset. This can cause
length issues as there may be variables that are excluded from the
model for different reasons.}
}
\value{
vector
}
\description{
Uses the model to extract the outcome variable. Throws
error if unable to find the outcome.
}
\keyword{internal}
