% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prGetModelData}
\alias{prGetModelData}
\title{Get model data.frame}
\usage{
prGetModelData(x)
}
\arguments{
\item{x}{The fitted model.}
}
\value{
data.frame

TODO: Check if this cannot be replaced by model.frame()
}
\description{
Combines the \code{\link{prExtractPredictorsFromModel}}
with the \code{\link{prExtractOutcomeFromModel}} to generate
a full model dataset
}
\author{
max
}

