% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{prMapVariable2Name}
\alias{prMapVariable2Name}
\title{A function that tries to resolve what variable corresponds to what row}
\usage{
prMapVariable2Name(var_names, available_names, data, force_match = TRUE)
}
\arguments{
\item{var_names}{The variable names that are saught after}

\item{available_names}{The names that are available to search through}

\item{data}{The data set that is saught after}

\item{force_match}{Whether all variables need to be identified or not.
E.g. you may only want to use some variables and already pruned the
\code{available_names} and therefore wont have matches. This is the
case when \code{\link{getCrudeAndAdjustedModelData}} has been used together
with the \code{var_select} argument.}
}
\value{
\code{list} Returns a list with each element has the corresponding
 variable name and a subsequent list with the parameters \code{no_rows}
 and \code{location} indiciting the number of rows corresponding to that
 element and where those rows are located. For factors the list also contains
 \code{lvls} and \code{no_lvls}.
}
\description{
As both the \code{\link{getCrudeAndAdjustedModelData}} and the
\code{\link{printCrudeAndAdjustedModel}} need to now exactly
what name from the \code{\link[stats]{coef}}/\code{\link[rms]{summary.rms}}
correspond to we for generalizeability this rather elaborate function.
}
\keyword{internal}
