% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/private_plotHR.R
\name{prPhRugPlot}
\alias{prPhRugPlot}
\title{Plot a rug on the datapoints}
\usage{
prPhRugPlot(xvalues)
}
\arguments{
\item{xvalues}{The xvalues that are used for the density}
}
\value{
\code{void}
}
\description{
Plot a rug on the datapoints
}
\keyword{internal}

